Fix Spotify's Notice at the beginning of the Chart
The previous `header_handler` method returned a `Dataframe` with Multindex columns. For example, `charts['Artist']` was a `DataFrame`, instead of `Series`. That didn’t allow to use methods such as `describe`, `unique` and `apply` functions in the normal way. This bug fix solves the header keeping the `DataFrame`'s usual properties.